Add a new bindings2/scripts directory for Dart bindings
These bindings were developed over the last 2 weeks on:
https://github.com/eseidel/skydart
They started as copies of the Dartium scripts (aka dart/tools/dom):
https://code.google.com/p/dart/source/browse/#svn%2Ftrunk%2Fdart%2Ftools%2Fdom
Which in turn started as a forked copy of Blink's
binding generation scripts from some years ago.

This is still way more python than we need to generate
bindings for SkyDart, but it's a start.

Reviews were done for many (but not all) of the patches.

I'll add the bindings2/ code and build system wiring
in further patches. Landing the new directories one at a
time so that it's easier to read through.

R=abarth@chromium.org, jamesr@chromium.org
BUG=454613

Review URL: https://codereview.chromium.org/914413004