Call DiscardableMemoryAllocator::SetInstance()
Apparently this can cause a crash when when we try to render an emoji at a
different size than the emoji font's native bitmap size. I'm not sure how to
reproduce the issue. I'll ask Hixie when he gets back.

Currently we just use regular memory. In the future, we should actually use
discarable memory.

Fixes #203