[Impeller] fix colr/bitmap font color drawing. (#52871)
Fixes https://github.com/flutter/flutter/issues/126546

Track the paint color used by Bitmap/Emoji/COLR fonts and use it as a cache key. This allows non-COLR glyphs in a COLR font to get the correct text color. For other kinds of fonts, we always record the color as black so there are no cache efficiency hits in general.