[CP-stable]Add xcprivacy privacy manifest to macOS framework (#55366)
### Issue Link:

https://github.com/flutter/flutter/issues/154915

### Changelog Description:

Comply with the new Apple privacy manifest policy for the macOS Flutter engine framework and prevent the "Missing privacy manifest" warning when submitting a macOS app to the App Store.

### Impact Description:

The macOS App Store will start enforcing privacy manifests for third-party SDKs starting November 12, 2024.  This change adds an empty privacy manifest to the Flutter engine FlutterMacOS.framework to comply with this new policy and prevents the warning.

### Workaround:
Add an empty PrivacyInfo.xcprivacy plist to FlutterMacOS.framework and re-codesign it manually.

### Risk:
What is the risk level of this cherry-pick?

### Test Coverage:
Are you confident that your fix is well-tested by automated tests?

A test was added to the framework, which should also be cherry-picked https://github.com/flutter/flutter/pull/155556.
While we can confirm the expected file is present, only a TestFlight/App Store submission will confirm the warning is gone.

### Validation Steps:
Submit a Flutter macOS app to TestFlight or the App Store.  Confirm there is no submission warning email sent.